app/server/ics_dc_detector/requirements.txt
# ── Elasticsearch ──────────────────────────────────────────────────────────────
elasticsearch[async]>=8.12,<9.0
elastic-transport>=8.12

# ── Matching engines ───────────────────────────────────────────────────────────
pyahocorasick>=2.0.0          # Stage 1 – Aho-Corasick trie
flashtext>=2.7                 # Stage 1 fallback / alias
regex>=2023.12.25              # Stage 2 – advanced regex (hyperscan-style batch)
sentence-transformers>=2.7.0   # Stage 3 – SecureBERT / cybersec embeddings
torch>=2.0.0                   # backend for sentence-transformers
numpy>=1.26.0
scikit-learn>=1.4.0            # cosine similarity helpers

# ── Async / Runtime ────────────────────────────────────────────────────────────
aiohttp>=3.9.0
aiofiles>=23.2.1
asyncio-throttle>=1.0.2

# ── State backends ─────────────────────────────────────────────────────────────
redis>=5.0.0
aioredis>=2.0.1

# ── Output sinks ───────────────────────────────────────────────────────────────
aiokafka>=0.10.0              # Kafka producer
aiohttp>=3.9.0                # HTTP webhook

# ── Observability ──────────────────────────────────────────────────────────────
prometheus-client>=0.20.0
structlog>=24.1.0

# ── Config & CLI ───────────────────────────────────────────────────────────────
pydantic>=2.6.0
pydantic-settings>=2.2.0
click>=8.1.7
python-dotenv>=1.0.1
rich>=13.7.0                  # pretty console output

# ── Utilities ──────────────────────────────────────────────────────────────────
tenacity>=8.2.3               # retry / back-off
orjson>=3.9.15                # fast JSON serialisation
python-dateutil>=2.9.0
cachetools>=5.3.3
